버그 수정 완료 02.16 [프로젝트 완성]

+++ src/controllers/boardController.js
@@ -1,5 +1,6 @@
 import { render } from "pug";
 import Board from "../models/Board";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   const { id } = req.params;
@@ -22,31 +23,44 @@ export const search = async (req, res) => {
 
 export const getSeeBoard = async (req, res) => {
   const { id } = req.params;
-  const board = await Board.findById(id);
+  const board = await Board.findById(id).populate("owner");
   if (!board) {
     return res.render("404", { pageTitle: "Page Not Found" });
   }
+  console.log(board);
   return res.render("watch", { board, pageTitle: "Watch" });
 };
 
 export const getEditBoard = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const board = await Board.findById(id);
   if (!board) {
-    return res.render("404", { pageTitle: "Page Not Found" });
+    return res.status(404).render("404", { pageTitle: "Page Not Found" });
+  }
+  if (String(board.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { board, pageTitle: "Edit" });
 };
 export const postEditBoard = async (req, res) => {
   const {
+    user: { _id },
+  } = req.session;
+  const {
     params: { id },
-    body: { title, content,boardImg },
+    body: { title, content, boardImg },
     file,
   } = req;
 
   const board = await Board.findById(id);
   if (!board) {
-    return res.render("404", { pageTitle: "Page Not Found" });
+    return res.status(404).render("404", { pageTitle: "Page Not Found" });
+  }
+  if (String(board.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Board.findByIdAndUpdate(id, {
     title,
@@ -62,15 +76,22 @@ export const getWriteBoard = (req, res) => {
 
 export const postWriteBoard = async (req, res) => {
   const {
-    body: { title, content,baordImg },
+    user: { _id },
+  } = req.session;
+  const {
+    body: { title, content, baordImg },
     file,
   } = req;
   try {
-    await Board.create({
+    const newBoard = await Board.create({
       title,
       content,
+      owner: _id,
       boardImg: file ? file.path : baordImg,
     });
+    const user = await User.findById(_id);
+    user.boards.push(newBoard._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("write", {
@@ -82,6 +103,19 @@ export const postWriteBoard = async (req, res) => {
 
 export const getDeleteBoard = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
+  const board = await Board.findById(id);
+  if (!board) {
+    return res.status(404).render("404", { pageTitle: "Page Not Found" });
+  }
+  if (String(board.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Board.findByIdAndDelete(id);
+
   return res.redirect("/");
 };
